+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+## v0.9.0
 - Folder assets are now marked with the `folder` attribute. The path is defined as for any other asset in `path`.
   - This adds support for dynamic folders in asset collections ([#28](https://github.com/NiklasEi/bevy_asset_loader/issues/28))
 - Support optional dynamic assets in collections ([#24](https://github.com/NiklasEi/bevy_asset_loader/issues/24))